records: moving 'views.py' as legacy code
* TOREMOVE after needed bits of code are moved
